+++ src/ast/css_nested_at.cr
@@ -7,8 +7,8 @@ module Mint
                      @content : String,
                      @name : String,
                      @file : Parser::File,
-                     @from : Int32,
-                     @to : Int32)
+                     @from : Int64,
+                     @to : Int64)
       end
     end
   end
