+++ src/ast/env.cr
@@ -5,8 +5,8 @@ module Mint
 
       def initialize(@name : String,
                      @file : Parser::File,
-                     @from : Int32,
-                     @to : Int32)
+                     @from : Int64,
+                     @to : Int64)
       end
     end
   end
