+++ src/ast/node.cr
@@ -2,7 +2,7 @@ module Mint
   class Ast
     class Node
       # Line and column number pair.
-      alias Position = {Int32, Int32}
+      alias Position = {Int64, Int64}
 
       struct Location
         include JSON::Serializable
@@ -22,11 +22,11 @@ module Mint
         def initialize(@filename, @start, @end)
         end
 
-        def contains?(line : Int)
+        def contains?(line : Int64)
           start[0] <= line <= end[0]
         end
 
-        def contains?(line : Int, column : Int)
+        def contains?(line : Int64, column : Int64)
           case
           when line == start[0] == end[0] # If on the only line
             start[1] <= column < end[1]
@@ -42,8 +42,8 @@ module Mint
 
       getter file : Parser::File
 
-      property from : Int32
-      getter to : Int32
+      property from : Int64
+      getter to : Int64
 
       def initialize(@file, @from, @to)
       end
@@ -74,8 +74,8 @@ module Mint
 
       def self.compute_position(lines, needle) : Position
         line_start_pos, line = begin
-          left, right = 0, lines.size - 1
-          index = pos = 0
+          left, right = 0_i64, lines.size - 1_i64
+          index = pos = 0_i64
           found = false
 
           while left <= right
@@ -83,9 +83,9 @@ module Mint
 
             case pos = lines[middle]
             when .< needle
-              left = middle + 1
+              left = middle + 1_i64
             when .> needle
-              right = middle - 1
+              right = middle - 1_i64
             else
               index = middle
               found = true
@@ -94,7 +94,7 @@ module Mint
           end
 
           unless found
-            index = left - 1
+            index = left - 1_i64
             pos = lines[index]
           end
 
@@ -102,7 +102,7 @@ module Mint
         end
 
         # NOTE: for the line numbers use 1-based indexing
-        line += 1
+        line += 1_i64
         column = needle - line_start_pos
 
         {line, column}
