+++ src/ast/operation.cr
@@ -7,8 +7,8 @@ module Mint
                      @left : Expression,
                      @operator : String,
                      @file : Parser::File,
-                     @from : Int32,
-                     @to : Int32)
+                     @from : Int64,
+                     @to : Int64)
       end
     end
   end
